+++ modules/hanna-utils/src/_/printDate.ts
@@ -1,3 +1,5 @@
+import { DEFAULT_LANG } from '../i18n';
+
 import { OpenRecord } from './types';
 
 const monthsByLang: OpenRecord<'is' | 'en' | 'pl', Array<string>> = {
@@ -48,14 +50,18 @@ const monthsByLang: OpenRecord<'is' | 'en' | 'pl', Array<string>> = {
 };
 
 /**
- * Very simple, very stupid, standalone date formatter for Icelandic, English and Polish.
+ * Very simple, very stupid, standalone date formatter for Icelandic (default),
+ * English and Polish.
  *
- * Just prints the full date (day, month, year).
+ * Just prints the full date (day, month, year). No bells. No whistles.
+ * No options.
+ *
+ * @see https://www.npmjs.com/package/@reykjavik/hanna-utils#printdate
  */
 export const printDate = (date: string | Date, lang?: string): string => {
   date = typeof date === 'string' ? new Date(date) : date;
   const d = date.getUTCDate();
-  const months = monthsByLang[lang || ''] || monthsByLang.en;
+  const months = monthsByLang[lang || ''] || monthsByLang[DEFAULT_LANG];
   const mmm = months[date.getUTCMonth()];
   const yyyy = date.getUTCFullYear();
   return months === monthsByLang.en ? `${mmm} ${d}, ${yyyy}` : `${d}. ${mmm} ${yyyy}`;