+++ modules/hanna-utils/src/index.ts
@@ -7,4 +7,5 @@ export * from './_/getSVGtext';
 export * from './_/misc';
 export * from './_/ObjectHelpers';
 export * from './_/printDate';
+export * from './_/testing';
 export * from './_/types';