+++ modules/hanna-utils/src/shareButtonsUtils.ts
@@ -2,6 +2,7 @@ import addUrlParams from '@hugsmidjan/qj/addUrlParams';
 import htmlLang from '@hugsmidjan/qj/htmlLang';
 
 import { ObjectKeys } from './_/ObjectHelpers';
+import { DEFAULT_LANG } from './i18n';
 
 const shareButtonPlatforms = {
   facebook: 'Facebook',
@@ -9,10 +10,19 @@ const shareButtonPlatforms = {
   linkedin: 'LinkedIn',
   email: 'E-mail',
 };
+/**
+ * List of supported social media platforms
+ */
 export type ShareButtonPlatforms = keyof typeof shareButtonPlatforms;
 
+/**
+ * List of supported social media platforms
+ */
 export const shareButtonTypes = ObjectKeys(shareButtonPlatforms);
 
+/**
+ * Texts for the social media sharing UI
+ */
 export type ShareButtonI18n = {
   label?: string;
   buttonLabel?: string;
@@ -24,6 +34,14 @@ export type ShareButtonI18n = {
 let _lastLabel: string | undefined;
 let _lastValue: [string, string] = ['', ''];
 const _token = '${name}';
+
+/**
+ * Generates a sharing button/link label for a given social media
+ * platform type.
+ *
+ * Interpolates the platform name, if the provided label contains
+ * a `${name}` token.
+ */
 export const getShareButtonLabel = (
   type: ShareButtonPlatforms,
   label: string
@@ -62,22 +80,36 @@ const getAttr = (selector: string, prop: string) => {
 
 // ---------------------------------------------------------------------------
 
-export const i18n: {
+/**
+ * Translation strings for social media sharing
+ */
+export const shareButtonTexts: {
   en: ShareButtonI18n;
+  is: ShareButtonI18n;
+  pl: ShareButtonI18n;
   [x: string]: ShareButtonI18n;
 } = {
-  en: {
-    label: 'Share this page via',
-    // buttonLabel: '',
-    emailSubject: 'Kíktu á þetta!',
-  },
   is: {
     label: 'Deila síðu á',
     // buttonLabel: '',
+    emailSubject: 'Kíktu á þetta!',
+  },
+  en: {
+    label: 'Share this page via',
+    // buttonLabel: '',
     emailSubject: 'Check this out!',
+  },
+  pl: {
+    label: 'Udostępnij tę stronę przez',
+    // buttonLabel: '',
+    emailSubject: 'Spójrz na to!',
   },
 };
 
+/**
+ * MetaData for the current page, including fully constructed sharing
+ * URLs for all `ShareButtonPlatforms`
+ */
 export type DocMeta = {
   hrefs: Record<ShareButtonPlatforms, string>;
   url: string;
@@ -88,11 +120,18 @@ export type DocMeta = {
 
 type DocMetaConfig = {
   emailSubject?: string;
+  lang?: string;
 };
 
-export const getDocMeta = (cfg?: DocMetaConfig): DocMeta => {
-  const { emailSubject } = cfg || {};
-
+/**
+ * Runs in the browser and parses the open-graph <meta/> tags
+ * of the current page.
+ *
+ * Returns the canonical `url`, `title`, `description` and page `lang`
+ * as well as ready-made "share this page" URLs for all
+ * `ShareButtonPlatforms`.
+ */
+export const getDocMeta = (cfg: DocMetaConfig = {}): DocMeta => {
   const url =
     getElm<HTMLLinkElement>('link[rel="canonical"]').href ||
     document.location.href.split('#')[0]!.replace(/[?&]fb_action_ids=.+/, '');
@@ -104,7 +143,7 @@ export const getDocMeta = (cfg?: DocMetaConfig): DocMeta => {
     getAttr('meta[property="og:description"]', 'content') ||
     getAttr('meta[name="description"]', 'content') ||
     '';
-  const lang = htmlLang() || '';
+  const lang = cfg.lang || htmlLang() || '';
 
   return {
     hrefs: {
@@ -121,7 +160,9 @@ export const getDocMeta = (cfg?: DocMetaConfig): DocMeta => {
         summary: description,
       }),
       email: addUrlParams('mailto:', {
-        subject: emailSubject || (i18n[lang] || i18n.en).emailSubject,
+        subject:
+          cfg.emailSubject ||
+          (shareButtonTexts[lang] || shareButtonTexts[DEFAULT_LANG]).emailSubject,
         body: title + '\n' + url + '\n\n',
       }),
     },
@@ -132,6 +173,10 @@ export const getDocMeta = (cfg?: DocMetaConfig): DocMeta => {
   };
 };
 
+/**
+ * An event handler, ready to be assigned to sharing links (with `href=""`s)
+ * that opens the link in a small popup window.
+ */
 export const openInPopup = (e: { target: EventTarget; preventDefault: () => void }) => {
   if (e.target instanceof HTMLAnchorElement) {
     e.preventDefault();