+++ modules/hanna-utils/CHANGELOG.md
@@ -3,9 +3,27 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+
+## 0.1.16
+
+_2022-11-03_
+
+- docs: Improve JSDocs and add missing chapters to README
+- fix: `printDate` should default to DEFAULT_LANGUAGE, not English
+- `shareButtonUtils`:
+  - fix: Add support for Polish
+  - fix: Make Icelandic the default/fallback locale
+  - fix: Swap in Icelandic and English `emailSubject`s
+
+## 0.1.15
+
+_2022-10-25_
+
 - feat: Add `styleServerUrl` and `setStyleServerUrl()`
 - feat: Add `ensurePosInt` and branded `PositiveInteger` type
 - feat: Add `notFalsy` type guarding filter function
+- feat: Add illustration token `vegvisir`
+- feat(ts): Add testing types `Expect`, `Equals`, `Extends`, `NotExtends`
 
 ## 0.1.14